Factor out notification concerns into a NotificationComponent
Presently, the way we're handling notifications doesn't quite align with the Carbon [notifications pattern] ().

Carbon offers two main variants of notifications: inline notifications and toasts. What we're referring to as toast notifications are inline notifications in the Carbon sense. We don't currently have a way to represent toasts.

To remedy this, I've factored out the current toast notification implementation into a generic `NotificationComponent` that provides a type `prop` with which one can specify whether they want to show an inline notification or a toast.

Though their implementations differ, there are enough functional similarities between the two notification types to merit grouping them in a single component.